class3hw: include "hello.h" in greeting.c
One should always do that, so that the compiler can check the declaration
(in hello.h) of the greeting() functionfor consistency with its
definition (in greeting.c).

In particular, this now leads to a compiler error because it's not
consistent, but a compiler error is better (because it's specific) than
the random crash we had before.