class11: add .at() member function for vector
This serves as a way to demonstrate how one now doesn't need to use
a VEC macro anymore.

Two versions of .at()  are implemented -- one which is "const", ie., can
not actually modify the vector, so that'll just return the value at
a given index.

On the other hand, if the vector can be modified, we'll return a
reference, so assigning something to that reference will actually
update the value in the array.