class14hw: large problem, don't write, add timing
In order to get usable timing measurements, we need to use a much bigger problem.

At that point, we don't really want to write out all that data, so I commented out that part. I added timing use Wtime(), and ran the averaging 5 times in a row to see whether we get consistent numbers.