class22: mv calc_rhs into heat_eqn.h
I've chosen to put the whole function definition into the header, not just
a declaration and add the definition in a separate source file. It's not
necessary here, but this method is often used in modern C++ that relies
on templates, in particular when using xtensor it's often helpful to do this.